rename: restaurant 하위 폴더에 menu, restaurant 폴더 생성
