+++ src/main/java/com/unknown/deliveryserver/domain/order/application/OrderServiceImpl.java
@@ -13,12 +13,12 @@ import com.unknown.deliveryserver.domain.order.entity.Order;
 import com.unknown.deliveryserver.domain.order.entity.OrderItemOptions;
 import com.unknown.deliveryserver.domain.order.entity.OrderItems;
 import com.unknown.deliveryserver.domain.order.enumerated.OrderStatus;
-import com.unknown.deliveryserver.domain.restaurant.application.RestaurantService;
-import com.unknown.deliveryserver.domain.restaurant.dao.menu.MenuOptionDetailRepository;
-import com.unknown.deliveryserver.domain.restaurant.dao.menu.MenuRepository;
-import com.unknown.deliveryserver.domain.restaurant.entity.Restaurant;
-import com.unknown.deliveryserver.domain.restaurant.entity.menu.Menu;
-import com.unknown.deliveryserver.domain.restaurant.entity.menu.MenuOptionDetail;
+import com.unknown.deliveryserver.domain.restaurant.restaurant.application.RestaurantService;
+import com.unknown.deliveryserver.domain.restaurant.menu.dao.MenuOptionDetailRepository;
+import com.unknown.deliveryserver.domain.restaurant.menu.dao.MenuRepository;
+import com.unknown.deliveryserver.domain.restaurant.restaurant.entity.Restaurant;
+import com.unknown.deliveryserver.domain.restaurant.menu.entity.Menu;
+import com.unknown.deliveryserver.domain.restaurant.menu.entity.MenuOptionDetail;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
