--- src/main/java/com/unknown/deliveryserver/domain/restaurant/dao/menu/MenuOptionRepository.java
@@ -1,7 +1,0 @@
-package com.unknown.deliveryserver.domain.restaurant.dao.menu;
-
-import com.unknown.deliveryserver.domain.restaurant.entity.menu.MenuOption;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface MenuOptionRepository extends JpaRepository<MenuOption, Long> {
-}
