+++ src/main/java/com/unknown/deliveryserver/domain/restaurant/menu/dao/MenuOptionRepository.java
@@ -1,0 +1,7 @@
+package com.unknown.deliveryserver.domain.restaurant.menu.dao;
+
+import com.unknown.deliveryserver.domain.restaurant.menu.entity.MenuOption;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface MenuOptionRepository extends JpaRepository<MenuOption, Long> {
+}
